+++ R/augment_rows.R
@@ -495,7 +495,8 @@ add_footer_row <- function(x, top = TRUE, values = character(0), colwidths = int
 #' be used to adding titles on the top rows of the flextable.
 #'
 #' @param x a `flextable` object
-#' @param values a character vector, each element will
+#' @param values a character vector or a call to [as_paragraph()]
+#' to get formated content, each element will
 #' be added as a new row.
 #' @param top should the row be inserted at the top
 #' or the bottom. Default to TRUE.
@@ -504,23 +505,51 @@ add_footer_row <- function(x, top = TRUE, values = character(0), colwidths = int
 #' @section Illustrations:
 #' \if{html}{\figure{fig_add_header_lines_1.png}{options: width="300"}}
 #' @examples
+#' # ex 1----
 #' ft_1 <- flextable(head(iris))
 #' ft_1 <- add_header_lines(ft_1, values = "blah blah")
 #' ft_1 <- add_header_lines(ft_1, values = c("blah 1", "blah 2"))
 #' ft_1 <- autofit(ft_1)
 #' ft_1
+#'
+#' # ex 2----
+#' ft01 <- fp_text_default(color = "red")
+#' ft02 <- fp_text_default(color = "orange")
+#' ref <- c("(1)", "(2)")
+#' pars <- as_paragraph(
+#'   as_chunk(ref, props = ft02), " ",
+#'   as_chunk(rep("My tailor is rich", length(ref)), props = ft01)
+#' )
+#'
+#' ft_2 <- flextable(head(mtcars))
+#' ft_2 <- add_header_lines(ft_2, values = pars, top = FALSE)
+#' ft_2 <- add_header_lines(ft_2, values = ref, top = TRUE)
+#' ft_2 <- add_footer_lines(ft_2, values = "blah", top = TRUE)
+#' ft_2 <- add_footer_lines(ft_2, values = pars, top = TRUE)
+#' ft_2 <- add_footer_lines(ft_2, values = ref, top = FALSE)
+#' ft_2 <- autofit(ft_2)
+#' ft_2
 add_header_lines <- function(x, values = character(0), top = TRUE) {
   if (!inherits(x, "flextable")) {
     stop(sprintf("Function `%s` supports only flextable objects.", "add_header_lines()"))
   }
-
   values_map <- values
-
   if (top) values_map <- rev(values_map)
 
-  for (value in values_map) {
-    x <- add_header_row(x, values = value, colwidths = length(x$col_keys), top = top)
+  if(inherits(values, "paragraph")) {
+    for (ivalue in seq_len(length(values_map))) {
+      x <- add_header_row(x, values = "", colwidths = length(x$col_keys), top = top)
+      if(top) i <- 1
+      else i <- nrow_part(x, "header")
+      x <- mk_par(x, i = i, j = 1, value = values_map[ivalue], part = "header")
+    }
+  } else {
+    for (value in values_map) {
+      x <- add_header_row(x, values = value, colwidths = length(x$col_keys), top = top)
+    }
   }
+
+
   x
 }
 
@@ -549,9 +578,25 @@ add_footer_lines <- function(x, values = character(0), top = FALSE) {
     stop(sprintf("Function `%s` supports only flextable objects.", "add_footer_lines()"))
   }
 
-  for (value in values) {
-    x <- add_footer_row(x, values = value, colwidths = length(x$col_keys), top = top)
+  ivalues <- seq_len(length(values))
+  if (top) {
+    ivalues <- rev(ivalues)
   }
+
+  if(inherits(values, "paragraph")) {
+    for (ivalue in ivalues) {
+      x <- add_footer_row(x, values = "", colwidths = length(x$col_keys), top = top)
+      if(top) i <- 1
+      else i <- nrow_part(x, "footer")
+      x <- mk_par(x, i = i, j = 1, value = values[ivalue], part = "footer")
+    }
+  } else {
+    for (ivalue in ivalues) {
+      x <- add_footer_row(x, values = values[ivalue],
+                          colwidths = length(x$col_keys), top = top)
+    }
+  }
+
   x
 }