+++ NEWS.md
@@ -15,6 +15,8 @@ supported with one exception `ft.htmlscroll`.
 This allows less interaction with the 'R Markdown' or 'Quarto'
 eco-system and let to define it globally with `set_flextable_defaults()`.
 - new function `grid_chunk()` to let users add grid graphics
+- functions `add_header_lines()` and `add_footer_lines()` 
+now supports formatted paragraph made with `as_paragraph()`.
 
 ## Issues